Changed id data type to Long. Added setter for id. Created two GET and one POST in ItemController. Updated ItemRepository to reflect id datatype change.

+++ back-end/src/main/java/org/launchcode/backend/models/data/ItemRepository.java
@@ -9,7 +9,7 @@ import javax.transaction.Transactional;
 
 @Repository
 //@Transactional
-public interface ItemRepository extends CrudRepository <Item, Integer>{
+public interface ItemRepository extends CrudRepository <Item, Long>{
 
 
 }
